+++ test/core-options-verify/core.verifyoptions.localPath.test.js
@@ -1,5 +1,6 @@
 const util = require('util');
 const core = require("../../lib/core-options.js");
+const conf = require('../../index.conf.js')
 
 test("verifyOptionsLocalPath", () => {
     var workingObject = {
@@ -31,7 +32,7 @@ test("verifyOptionsLocalPath-blank", () => {
 
     var warnOutput = [];
     global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s); })
-    var msg = util.format("Warning: localPath undefined, defaulting to current directory. [%s]", ".");
+    var msg = util.format("Warning: localPath undefined, defaulting to current directory. [%s]", conf.DEFAULT_LOCALPATH);
 
     expect(core.verifyOptionsLocalPath(workingObject))
         .resolves
@@ -40,7 +41,7 @@ test("verifyOptionsLocalPath-blank", () => {
                 localPath: ""
             },
             validatedOption: {
-                localPath: "."
+                localPath: conf.DEFAULT_LOCALPATH
             }
         });
     expect(console.warn).toBeCalled();
@@ -54,13 +55,13 @@ test("verifyOptionsLocalPath-undefined", () => {
 
     var warnOutput = [];
     global.console.warn = jest.fn().mockImplementation((s) => { warnOutput.push(s); })
-    var msg = util.format("Warning: localPath undefined, defaulting to current directory. [%s]", ".");
+    var msg = util.format("Warning: localPath undefined, defaulting to current directory. [%s]", conf.DEFAULT_LOCALPATH);
     expect(core.verifyOptionsLocalPath(workingObject))
         .resolves
         .toMatchObject({
             userOption: {},
             validatedOption: {
-                localPath: "."
+                localPath: conf.DEFAULT_LOCALPATH
             }
         });
     expect(console.warn).toBeCalled();