+++ test/core-options-verify/core.verifyoptions.quiet.test.js
@@ -1,5 +1,6 @@
 const util = require('util');
 const core = require("../../lib/core-options.js");
+const conf = require('../../index.conf.js')
 
 test.each([true, "Y", "on", 1, "y", "yes"])("verifyOptionsQuiet-Good-True", (value) => {
     var workingObject = {
@@ -52,7 +53,7 @@ test("verifyOptionsQuiet-undefined", () => {
         .toMatchObject({
             userOption: {},
             validatedOption: {
-                quiet: false
+                quiet: conf.DEFAULT_QUIET
             }
         });
 });