+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## `0.10.0` upcoming
 
-- `#30`: As dev/test, I want to use jest for test framework. (Major code refactoring)
+- `#30`: As dev/test, I want to use jest for test framework.
+- `#43`: As dev, I want to move all default value to conf.js.
 
 ## `0.5.0`